requirements.txt
streamlit
Pillow
scikit-learn
numpy
streamlit-image-coordinates
